Add a SignalHandler for SIGINT (Ctrl-C)
Problem
Currently backup can be interrupted right after uploading an
archive and right before writing the status file to disk, leading
to a lone file in glacier without a local trace.

Solution
This handler properly cancel any task in progress.
It's still possible to have a crash at the wrong moment but it's less
likely that this problem will happened.